requirements.txt
django==5.1.1
djangorestframework
djangorestframework-simplejwt
dj-database-url
psycopg2-binary
django-environ
whitenoise
drf-yasg
